# docs/swagger.yaml
definitions:
  handler.CreateOpeningRequest:
    properties:
      company:
        type: string
      link:
        type: string
      location:
        type: string
      remote:
        type: boolean
      role:
        type: string
      salary:
        type: integer
    type: object
  handler.CreateOpeningResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.OpeningResponse'
      message:
        type: string
    type: object
  handler.DeleteOpeningResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.OpeningResponse'
      message:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      errorCode:
        type: string
      message:
        type: string
    type: object
  handler.ShowOpeningResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.OpeningResponse'
      message:
        type: string
    type: object
  schemas.OpeningResponse:
    properties:
      company:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      link:
        type: string
      location:
        type: string
      remote:
        type: boolean
      role:
        type: string
      salary:
        type: integer
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /opening:
    delete:
      consumes:
      - application/json
      description: Delete a opening
      parameters:
      - description: Opening identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.DeleteOpeningResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete Opening
      tags:
      - Openings
    get:
      description: Show a opening
      parameters:
      - description: Opening identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ShowOpeningResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Show Opening
      tags:
      - Openings
    post:
      consumes:
      - application/json
      description: Create a new job opening
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateOpeningRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CreateOpeningResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create Opening
      tags:
      - Openings
swagger: "2.0"